winit-core: move `ActiveEventLoop`
